+++ bin-src/lib/builders.ts
@@ -1,0 +1,5 @@
+export const builders = {
+  webpack4: '@storybook/builder-webpack4',
+  webpack5: '@storybook/builder-webpack5',
+  '@storybook/vite-builder': '@storybook/builder-vite',
+} as Record<string, string>;
