+++ bin-src/lib/getStorybookMetadata.ts
@@ -1,40 +1,185 @@
-import { readFile } from 'jsonfile';
+import fs from 'fs-extra';
+import meow from 'meow';
+import { parseArgsStringToArgv } from 'string-argv';
+import semver from 'semver';
+import { loadMainConfig } from '@storybook/core-common';
 import { Context } from '../types';
-import { supportedAddons } from './supportedAddons';
+import noViewLayerPackage from '../ui/messages/errors/noViewLayerPackage';
 import { viewLayers } from './viewLayers';
+import { timeout, raceFulfilled } from './promises';
 
-/* 
-  In Storybook 6.5+, when building a Storybook, a project.json file is generated,
-  containing metadata (sb builder, addons, viewLayer and viewLayer version) for the 
-  particular setup.
-*/
+const resolvePackageJson = (pkg: string) => {
+  try {
+    // we bundle this app for node, meaning all require calls are replaced by webpack.
+    // in this case we want to use node's actual require functionality!
+    // webpack will provide a '__non_webpack_require__' function to do this with,
+    // but this will obviously not be present during tests, hence the check and fallback to the normal require
+    const r = typeof __non_webpack_require__ !== 'undefined' ? __non_webpack_require__ : require;
+    const packagePath = r.resolve(`${pkg}/package.json`);
+    return fs.readJson(packagePath);
+  } catch (error) {
+    return Promise.reject(error);
+  }
+};
 
-export interface SBProjectJson {
-  addons: Record<string, { version: string; options: any }>;
-  builder?: {
-    name: string;
-  };
-  framework: {
-    name: string;
-  };
-  storybookPackages?: Record<string, { version: string }>;
-}
+const findDependency = (
+  { dependencies, devDependencies, peerDependencies },
+  predicate: Parameters<typeof Array.prototype.find>[0]
+) => [
+  Object.entries(dependencies || {}).find(predicate),
+  Object.entries(devDependencies || {}).find(predicate),
+  Object.entries(peerDependencies || {}).find(predicate),
+];
 
-export const getStorybookMetadateFromProjectJson = async (
-  projectJsonPath: string
-): Promise<Partial<Context['storybook']>> => {
-  const sbProjectJson = (await readFile(projectJsonPath)) as SBProjectJson;
-  const viewLayerPackage = Object.keys(viewLayers).find(
-    (viewLayer) => viewLayers[viewLayer] === sbProjectJson.framework.name
-  );
+const getDependencyInfo = ({ packageJson, log }, dependencyMap: Record<string, string>) => {
+  const [dep, devDep, peerDep] = findDependency(packageJson, ([key]) => dependencyMap[key]);
+  const [pkg, version] = dep || devDep || peerDep || [];
+  const dependency = viewLayers[pkg];
+
+  if (dep && devDep && dep[0] === devDep[0]) {
+    log.warn(
+      `Found "${dep[0]}" in both "dependencies" and "devDependencies". This is probably a mistake.`
+    );
+  }
+  if (dep && peerDep && dep[0] === peerDep[0]) {
+    log.warn(
+      `Found "${dep[0]}" in both "dependencies" and "peerDependencies". This is probably a mistake.`
+    );
+  }
+  return { dependency, version, dependencyPackage: pkg };
+};
+
+const findViewlayer = async ({ env, log, options, packageJson }) => {
+  // Allow setting Storybook version via CHROMATIC_STORYBOOK_VERSION='@storybook/react@4.0-alpha.8' for unusual cases
+  if (env.CHROMATIC_STORYBOOK_VERSION) {
+    const [, p, v] = env.CHROMATIC_STORYBOOK_VERSION.match(/(.+)@(.+)$/) || [];
+    const version = semver.valid(v); // ensures we get a specific version, not a range
+    if (!p || !version) {
+      throw new Error(
+        'Invalid CHROMATIC_STORYBOOK_VERSION; expecting something like "@storybook/react@6.2.0".'
+      );
+    }
+    const viewLayer = viewLayers[p] || viewLayers[`@storybook/${p}`];
+    if (!viewLayer) {
+      throw new Error(`Unsupported viewlayer specified in CHROMATIC_STORYBOOK_VERSION: ${p}`);
+    }
+    return { version, viewLayer };
+  }
+
+  const {
+    dependency: viewLayer,
+    version,
+    dependencyPackage: pkg,
+  } = getDependencyInfo({ log, packageJson }, viewLayers);
+
+  if (viewLayer) {
+    if (options.storybookBuildDir) {
+      // If we aren't going to invoke the Storybook CLI later, we can exit early.
+      // Note that `version` can be a semver range in this case.
+      return { viewLayer, version };
+    }
+    // Verify that the viewlayer package is actually present in node_modules.
+    return Promise.race([
+      resolvePackageJson(pkg)
+        .then((json) => ({ viewLayer, version: json.version }))
+        .catch(() => Promise.reject(new Error(noViewLayerPackage(pkg)))),
+      timeout(10000),
+    ]);
+  }
+
+  if (!options.interactive) {
+    log.info(`No viewlayer package listed in dependencies. Checking transitive dependencies.`);
+  }
+
+  // We might have a transitive dependency (e.g. through `@nuxtjs/storybook` which depends on
+  // `@storybook/vue`). In this case we look for any viewlayer package present in node_modules,
+  // and return the first one we find.
+  return Promise.race([
+    raceFulfilled(
+      Object.entries(viewLayers).map(async ([key, value]) => {
+        const json = await resolvePackageJson(key);
+        return { viewLayer: value, version: json.version };
+      })
+    ).catch(() => Promise.reject(new Error(noViewLayerPackage(pkg)))),
+    timeout(10000),
+  ]);
+};
+
+const findAddons = async (deps, mainConfig) => {
+  if (mainConfig.addons) {
+    return {
+      addons: mainConfig.addons.map((addon) => {
+        let name: string;
+        let options: any;
+        if (typeof addon === 'string') {
+          name = addon.replace('/register', '');
+        } else {
+          options = addon.options;
+          name = addon.name;
+        }
+
+        return {
+          name,
+          options,
+          version: deps[name],
+        };
+      }),
+    };
+  }
   return {
-    viewLayer: sbProjectJson.framework.name ?? '',
-    version: sbProjectJson.storybookPackages[viewLayerPackage].version,
-    builder: sbProjectJson.builder.name ?? 'webpack4',
-    addons: Object.keys(sbProjectJson.addons).map((addon) => ({
-      name: supportedAddons[addon],
-      packageName: addon,
-      packageVersion: sbProjectJson.addons[addon].version,
-    })),
+    addons: [],
   };
+};
+
+const findConfigFlags = async ({ options, packageJson }) => {
+  const { scripts = {} } = packageJson;
+  if (!options.buildScriptName || !scripts[options.buildScriptName]) return {};
+
+  const { flags } = meow({
+    argv: parseArgsStringToArgv(scripts[options.buildScriptName]),
+    flags: {
+      configDir: { type: 'string', alias: 'c' },
+      staticDir: { type: 'string', alias: 's' },
+    },
+  });
+
+  return {
+    configDir: flags.configDir,
+    staticDir: flags.staticDir && flags.staticDir.split(','),
+  };
+};
+
+const findBuilder = async (mainConfig) => {
+  if (mainConfig.core?.builder) {
+    const { builder } = mainConfig.core;
+
+    return {
+      name: typeof builder === 'string' ? builder : builder.name,
+      options: typeof builder === 'string' ? undefined : builder?.options ?? undefined,
+    };
+  }
+  return {
+    builder: {
+      name: 'webpack4',
+      options: undefined,
+    },
+  };
+};
+
+export const getStorybookMetadata = async (ctx: Context) => {
+  const configDir = ctx.options.storybookConfigDir ?? '.storybook';
+  const mainConfig = loadMainConfig({ configDir });
+  const allDependencies = {
+    ...ctx.packageJson?.dependencies,
+    ...ctx.packageJson?.devDependencies,
+    ...ctx.packageJson?.peerDependencies,
+  };
+
+  const info = await Promise.all([
+    findAddons(allDependencies, mainConfig),
+    findConfigFlags(ctx),
+    findViewlayer(ctx),
+    findBuilder(mainConfig),
+  ]);
+  return info.reduce((acc, obj) => Object.assign(acc, obj), {});
 };
